Feat: 카테고리 삭제 로직 및 공부 시작 시 category 검증, 일부 모호한 이름 변경
- StudyCategory -> PersonalStudyCategory 변경 (dao 역시 엔티티와 동일하게)
- 카테고리 삭제 시 : studyTime 의 studyType 을 remove, typeId 를 null 로 변경 후 PersonalStudyCategory 삭제
- 위 삭제 로직을 위해 영속성 전략 패턴(persistenceStrategy)에서 알맞는 카테고리 정보를 불러오는 메서드 정의 및 구현
- 위 삭제 로직을 위해 study status 에 goal 등을 추가
- 위 삭제 로직을 위해 studyCategory 추상 클래스 정의 및 카테고리 관련 엔티티가 이를 상속받도록 함

- 공부 시작 시 모호한 값이 들어와 오류가 생기는 것을 방지하기 위한 매개변수 검증 처리 추가

Ref: #97

+++ src/main/java/com/studypals/domain/groupManage/dto/GetGroupCategoryRes.java
@@ -16,6 +16,6 @@ public record GetGroupCategoryRes(
         Long typeId,
         String name,
         String color,
-        Integer goalTime,
+        Long goalTime,
         Integer dayBelong,
         String description) {}
